feat(avatar): add support for multiple image sizes
When an avatar is uploaded, we will now generate multiple sizes of the
image, to better accommodate various environments that the avatar may
appear.

The sizes generated are based of current constants in the `lib/img/index.js`
module. The `SIZES` constant contains a map of `suffix` to `height/width`
values. The suffix `default` is special cased, and will actually mean no
suffix. Every other value will be appended on the end of the URL, with an
underscore.

The current `SIZES` are:

    {
      small: { h: 100, w: 100 },
      default: { h: 200, w: 200 },
      large: { h 600, w: 600 }
    }

which generate 3 images when an upload occurs:

- https://s3.url/<imageid>_small (100px x 100px)
- https://s3.url/<imageid> (200px x 200px)
- https://s3.url/<imageid>_large (600px x 600px)

Closes #68
Closes #89